+++ src/Reference.php
@@ -8,22 +8,31 @@ final class Reference {
     private static ?array $data = null;
 
     /**
+     * @return string[]
+     * @phpstan-return non-empty-array<string[]>
      * @throws \ErickSkrauch\MccMnc\UnknownException
      */
-    public static function countryFromMcc(int $mcc): string {
+    public static function countryFromMcc(int $mcc): array {
+        $countries = [];
         foreach (self::getData() as $data) {
-            if ($data['mcc'] === $mcc) {
-                return $data['code'];
+            if (in_array($mcc, $data['mcc'], true)) {
+                $countries[] = $data['code'];
             }
         }
 
-        throw new UnknownException();
+        if (empty($countries)) {
+            throw new UnknownException();
+        }
+
+        return $countries;
     }
 
     /**
+     * @return int[]
+     * @phpstan-return non-empty-array<int[]>
      * @throws \ErickSkrauch\MccMnc\UnknownException
      */
-    public static function mccFromCountry(string $countryCode): int {
+    public static function mccFromCountry(string $countryCode): array {
         foreach (self::getData() as $data) {
             if ($data['code'] === $countryCode) {
                 return $data['mcc'];