+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## 0.2.0 - 2023-05-01
+### Added
+- More MCCs.
+
+### Changed
+- `Reference::countryFromMcc()` now returns an array of countries, which belongs to provided MCC.
+- `Reference::mccFromCountry()` now returns an array of MCCs, which belongs to provided country code.
 
 ## 0.1.0 - 2023-04-25
 ### Added